+++ src/prefect/serialization/storage.py
@@ -1,8 +1,8 @@
 import marshmallow
-from marshmallow import fields
+from marshmallow import fields, post_load
 
 import prefect
-from prefect.environments.storage import Bytes, Docker, Storage
+from prefect.environments.storage import Memory, Docker, Storage
 from prefect.utilities.serialization import ObjectSchema, OneOfSchema
 
 
@@ -18,12 +18,19 @@ class DockerSchema(ObjectSchema):
     registry_url = fields.String(allow_none=True)
     image_name = fields.String(allow_none=True)
     image_tag = fields.String(allow_none=True)
-    flow_file_path = fields.String(allow_none=True)
+    flows = fields.Dict(key=fields.Str(), values=fields.Str())
+
+    @post_load
+    def create_object(self, data: dict) -> Docker:
+        flows = data.pop("flows", dict())
+        base_obj = super().create_object(data)
+        base_obj.flows = flows
+        return base_obj
 
 
-class BytesSchema(ObjectSchema):
+class MemorySchema(ObjectSchema):
     class Meta:
-        object_class = Bytes
+        object_class = Memory
 
 
 class StorageSchema(OneOfSchema):
@@ -34,6 +41,6 @@ class StorageSchema(OneOfSchema):
     # map class name to schema
     type_schemas = {
         "Docker": DockerSchema,
-        "Bytes": BytesSchema,
+        "Memory": MemorySchema,
         "Storage": BaseStorageSchema,
     }